Address comments by @trustin and @ikhoon
